backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
click==8.2.1
distro==1.9.0
exceptiongroup==1.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
joblib==1.5.1
nltk==3.9.1
openai==1.86.0
pydantic==2.11.5
pydantic_core==2.33.2
python-dotenv==1.1.0
regex==2024.11.6
sniffio==1.3.1
textblob==0.19.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
